+++ src/react/apis/string/IString.ts
@@ -1,0 +1,9 @@
+export interface IString {
+  id: string
+  name: string
+  order: number
+  thread: string
+  description?: string
+  dateCreated: string
+  dateModified: string
+}