+++ src/react/apis/thread/get-threads-api.ts
@@ -1,6 +1,6 @@
 import AxiosApi from '../AxiosApi'
 import { threadController } from '../../constants/urls'
-import { IThread } from './ithread'
+import { IThread } from './IThread'
 import { AxiosRequestConfig } from 'axios'
 
 export const getThreadsApi = new AxiosApi<AxiosRequestConfig, IThread[]>('GET',